added function and test for counting ground atoms for each timestep
