+++ pyreason/scripts/interpretation/interpretation_parallel.py
@@ -52,8 +52,6 @@ edges_to_be_added_type = numba.types.Tuple((numba.types.ListType(node_type), num
 
 
 class Interpretation:
-	available_labels_node = []
-	available_labels_edge = []
 	specific_node_labels = numba.typed.Dict.empty(key_type=label.label_type, value_type=numba.types.ListType(node_type))
 	specific_edge_labels = numba.typed.Dict.empty(key_type=label.label_type, value_type=numba.types.ListType(edge_type))
 
@@ -69,6 +67,10 @@ class Interpretation:
 		self.store_interpretation_changes = store_interpretation_changes
 		self.update_mode = update_mode
 		self.allow_ground_rules = allow_ground_rules
+
+		# Counter for number of ground atoms for each timestep, start with zero for the zeroth timestep
+		self.num_ga = numba.typed.List.empty_list(numba.types.int64)
+		self.num_ga.append(0)
 
 		# For reasoning and reasoning again (contains previous time and previous fp operation cnt)
 		self.time = 0
@@ -98,18 +100,8 @@ class Interpretation:
 		self.nodes.extend(numba.typed.List(self.graph.nodes()))
 		self.edges.extend(numba.typed.List(self.graph.edges()))
 
-		# Make sure they are correct type
-		if len(self.available_labels_node)==0:
-			self.available_labels_node = numba.typed.List.empty_list(label.label_type)
-		else:
-			self.available_labels_node = numba.typed.List(self.available_labels_node)
-		if len(self.available_labels_edge)==0:
-			self.available_labels_edge = numba.typed.List.empty_list(label.label_type)
-		else:
-			self.available_labels_edge = numba.typed.List(self.available_labels_edge)
-
-		self.interpretations_node, self.predicate_map_node = self._init_interpretations_node(self.nodes, self.available_labels_node, self.specific_node_labels)
-		self.interpretations_edge, self.predicate_map_edge = self._init_interpretations_edge(self.edges, self.available_labels_edge, self.specific_edge_labels)
+		self.interpretations_node, self.predicate_map_node = self._init_interpretations_node(self.nodes, self.specific_node_labels, self.num_ga)
+		self.interpretations_edge, self.predicate_map_edge = self._init_interpretations_edge(self.edges, self.specific_edge_labels, self.num_ga)
 
 		# Setup graph neighbors and reverse neighbors
 		self.neighbors = numba.typed.Dict.empty(key_type=node_type, value_type=numba.types.ListType(node_type))
@@ -138,19 +130,19 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True)
-	def _init_interpretations_node(nodes, available_labels, specific_labels):
+	def _init_interpretations_node(nodes, specific_labels, num_ga):
 		interpretations = numba.typed.Dict.empty(key_type=node_type, value_type=world.world_type)
 		predicate_map = numba.typed.Dict.empty(key_type=label.label_type, value_type=list_of_nodes)
-		# General labels
+
+		# Initialize nodes
 		for n in nodes:
-			interpretations[n] = world.World(available_labels)
+			interpretations[n] = world.World(numba.typed.List.empty_list(label.label_type))
+
 		# Specific labels
 		for l, ns in specific_labels.items():
 			for n in ns:
 				interpretations[n].world[l] = interval.closed(0.0, 1.0)
-
-		for l in available_labels:
-			predicate_map[l] = numba.typed.List(nodes)
+				num_ga[0] += 1
 
 		for l, ns in specific_labels.items():
 			predicate_map[l] = numba.typed.List(ns)
@@ -159,19 +151,19 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True)
-	def _init_interpretations_edge(edges, available_labels, specific_labels):
+	def _init_interpretations_edge(edges, specific_labels, num_ga):
 		interpretations = numba.typed.Dict.empty(key_type=edge_type, value_type=world.world_type)
 		predicate_map = numba.typed.Dict.empty(key_type=label.label_type, value_type=list_of_edges)
-		# General labels
-		for e in edges:
-			interpretations[e] = world.World(available_labels)
+
+		# Initialize edges
+		for n in edges:
+			interpretations[n] = world.World(numba.typed.List.empty_list(label.label_type))
+
 		# Specific labels
 		for l, es in specific_labels.items():
 			for e in es:
 				interpretations[e].world[l] = interval.closed(0.0, 1.0)
-
-		for l in available_labels:
-			predicate_map[l] = numba.typed.List(edges)
+				num_ga[0] += 1
 
 		for l, es in specific_labels.items():
 			predicate_map[l] = numba.typed.List(es)
@@ -221,7 +213,7 @@ class Interpretation:
 		return max_time
 
 	def _start_fp(self, rules, max_facts_time, verbose, again):
-		fp_cnt, t = self.reason(self.interpretations_node, self.interpretations_edge, self.predicate_map_node, self.predicate_map_edge, self.tmax, self.prev_reasoning_data, rules, self.nodes, self.edges, self.neighbors, self.reverse_neighbors, self.rules_to_be_applied_node, self.rules_to_be_applied_edge, self.edges_to_be_added_node_rule, self.edges_to_be_added_edge_rule, self.rules_to_be_applied_node_trace, self.rules_to_be_applied_edge_trace, self.facts_to_be_applied_node, self.facts_to_be_applied_edge, self.facts_to_be_applied_node_trace, self.facts_to_be_applied_edge_trace, self.ipl, self.rule_trace_node, self.rule_trace_edge, self.rule_trace_node_atoms, self.rule_trace_edge_atoms, self.reverse_graph, self.atom_trace, self.save_graph_attributes_to_rule_trace, self.canonical, self.inconsistency_check, self.store_interpretation_changes, self.update_mode, self.allow_ground_rules, max_facts_time, self.annotation_functions, self._convergence_mode, self._convergence_delta, verbose, again)
+		fp_cnt, t = self.reason(self.interpretations_node, self.interpretations_edge, self.predicate_map_node, self.predicate_map_edge, self.tmax, self.prev_reasoning_data, rules, self.nodes, self.edges, self.neighbors, self.reverse_neighbors, self.rules_to_be_applied_node, self.rules_to_be_applied_edge, self.edges_to_be_added_node_rule, self.edges_to_be_added_edge_rule, self.rules_to_be_applied_node_trace, self.rules_to_be_applied_edge_trace, self.facts_to_be_applied_node, self.facts_to_be_applied_edge, self.facts_to_be_applied_node_trace, self.facts_to_be_applied_edge_trace, self.ipl, self.rule_trace_node, self.rule_trace_edge, self.rule_trace_node_atoms, self.rule_trace_edge_atoms, self.reverse_graph, self.atom_trace, self.save_graph_attributes_to_rule_trace, self.canonical, self.inconsistency_check, self.store_interpretation_changes, self.update_mode, self.allow_ground_rules, max_facts_time, self.annotation_functions, self._convergence_mode, self._convergence_delta, self.num_ga, verbose, again)
 		self.time = t - 1
 		# If we need to reason again, store the next timestep to start from
 		self.prev_reasoning_data[0] = t
@@ -231,7 +223,7 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True, parallel=True)
-	def reason(interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, tmax, prev_reasoning_data, rules, nodes, edges, neighbors, reverse_neighbors, rules_to_be_applied_node, rules_to_be_applied_edge, edges_to_be_added_node_rule, edges_to_be_added_edge_rule, rules_to_be_applied_node_trace, rules_to_be_applied_edge_trace, facts_to_be_applied_node, facts_to_be_applied_edge, facts_to_be_applied_node_trace, facts_to_be_applied_edge_trace, ipl, rule_trace_node, rule_trace_edge, rule_trace_node_atoms, rule_trace_edge_atoms, reverse_graph, atom_trace, save_graph_attributes_to_rule_trace, canonical, inconsistency_check, store_interpretation_changes, update_mode, allow_ground_rules, max_facts_time, annotation_functions, convergence_mode, convergence_delta, verbose, again):
+	def reason(interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, tmax, prev_reasoning_data, rules, nodes, edges, neighbors, reverse_neighbors, rules_to_be_applied_node, rules_to_be_applied_edge, edges_to_be_added_node_rule, edges_to_be_added_edge_rule, rules_to_be_applied_node_trace, rules_to_be_applied_edge_trace, facts_to_be_applied_node, facts_to_be_applied_edge, facts_to_be_applied_node_trace, facts_to_be_applied_edge_trace, ipl, rule_trace_node, rule_trace_edge, rule_trace_node_atoms, rule_trace_edge_atoms, reverse_graph, atom_trace, save_graph_attributes_to_rule_trace, canonical, inconsistency_check, store_interpretation_changes, update_mode, allow_ground_rules, max_facts_time, annotation_functions, convergence_mode, convergence_delta, num_ga, verbose, again):
 		t = prev_reasoning_data[0]
 		fp_cnt = prev_reasoning_data[1]
 		max_rules_time = 0
@@ -302,7 +294,7 @@ class Interpretation:
 						if check_consistent_node(interpretations_node, comp, (l, bnd)):
 							mode = 'graph-attribute-fact' if graph_attribute else 'fact'
 							override = True if update_mode == 'override' else False
-							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode=mode, override=override)
+							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode=mode, override=override)
 
 							update = u or update
 							# Update convergence params
@@ -316,7 +308,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_node(interpretations_node, comp, (l, bnd), ipl, t, fp_cnt, i, atom_trace, rule_trace_node, rule_trace_node_atoms, rules_to_be_applied_node_trace, facts_to_be_applied_node_trace, store_interpretation_changes, mode=mode)
 							else:
-								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode=mode, override=True)
+								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode=mode, override=True)
 
 								update = u or update
 								# Update convergence params
@@ -344,7 +336,7 @@ class Interpretation:
 					comp, l, bnd, static, graph_attribute = facts_to_be_applied_edge[i][1], facts_to_be_applied_edge[i][2], facts_to_be_applied_edge[i][3], facts_to_be_applied_edge[i][4], facts_to_be_applied_edge[i][5]
 					# If the component is not in the graph, add it
 					if comp not in edges_set:
-						_add_edge(comp[0], comp[1], neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge)
+						_add_edge(comp[0], comp[1], neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
 						edges_set.add(comp)
 
 					# Check if bnd is static. Then no need to update, just add to rule trace, check if graph attribute, and add ipl complement to rule trace as well
@@ -368,7 +360,7 @@ class Interpretation:
 						if check_consistent_edge(interpretations_edge, comp, (l, bnd)):
 							mode = 'graph-attribute-fact' if graph_attribute else 'fact'
 							override = True if update_mode == 'override' else False
-							u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode=mode, override=override)
+							u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode=mode, override=override)
 
 							update = u or update
 							# Update convergence params
@@ -382,7 +374,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_edge(interpretations_edge, comp, (l, bnd), ipl, t, fp_cnt, i, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode=mode)
 							else:
-								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode=mode, override=True)
+								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode=mode, override=True)
 
 								update = u or update
 								# Update convergence params
@@ -416,7 +408,7 @@ class Interpretation:
 						# Check for inconsistencies
 						if check_consistent_node(interpretations_node, comp, (l, bnd)):
 							override = True if update_mode == 'override' else False
-							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode='rule', override=override)
+							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
 
 							update = u or update
 							# Update convergence params
@@ -429,7 +421,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_node(interpretations_node, comp, (l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_node, rule_trace_node_atoms, rules_to_be_applied_node_trace, facts_to_be_applied_node_trace, store_interpretation_changes, mode='rule')
 							else:
-								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode='rule', override=True)
+								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
 
 								update = u or update
 								# Update convergence params
@@ -453,7 +445,7 @@ class Interpretation:
 					if i[0] == t:
 						comp, l, bnd, immediate, set_static = i[1], i[2], i[3], i[4], i[5]
 						sources, targets, edge_l = edges_to_be_added_edge_rule[idx]
-						edges_added, changes = _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, edge_l, interpretations_node, interpretations_edge, predicate_map_edge)
+						edges_added, changes = _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, edge_l, interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
 						changes_cnt += changes
 
 						# Update bound for newly added edges. Use bnd to update all edges if label is specified, else use bnd to update normally
@@ -463,7 +455,7 @@ class Interpretation:
 									continue
 								if check_consistent_edge(interpretations_edge, e, (edge_l, bnd)):
 									override = True if update_mode == 'override' else False
-									u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=override)
+									u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
 
 									update = u or update
 
@@ -477,7 +469,7 @@ class Interpretation:
 									if inconsistency_check:
 										resolve_inconsistency_edge(interpretations_edge, e, (edge_l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode='rule')
 									else:
-										u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=True)
+										u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
 
 										update = u or update
 
@@ -491,7 +483,7 @@ class Interpretation:
 							# Check for inconsistencies
 							if check_consistent_edge(interpretations_edge, comp, (l, bnd)):
 								override = True if update_mode == 'override' else False
-								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=override)
+								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
 
 								update = u or update
 								# Update convergence params
@@ -504,7 +496,7 @@ class Interpretation:
 								if inconsistency_check:
 									resolve_inconsistency_edge(interpretations_edge, comp, (l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode='rule')
 								else:
-									u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=True)
+									u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
 
 									update = u or update
 									# Update convergence params
@@ -543,7 +535,7 @@ class Interpretation:
 						# Only go through if the rule can be applied within the given timesteps, or we're running until convergence
 						delta_t = rule.get_delta()
 						if t + delta_t <= tmax or tmax == -1 or again:
-							applicable_node_rules, applicable_edge_rules = _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules)
+							applicable_node_rules, applicable_edge_rules = _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules, num_ga, t)
 
 							# Loop through applicable rules and add them to the rules to be applied for later or next fp operation
 							for applicable_rule in applicable_node_rules:
@@ -630,14 +622,15 @@ class Interpretation:
 					t += 1
 					break
 
-			# Increment t
+			# Increment t, update number of ground atoms
 			t += 1
+			num_ga.append(num_ga[-1])
 
 		return fp_cnt, t
 
 	def add_edge(self, edge, l):
 		# This function is useful for pyreason gym, called externally
-		_add_edge(edge[0], edge[1], self.neighbors, self.reverse_neighbors, self.nodes, self.edges, l, self.interpretations_node, self.interpretations_edge, self.predicate_map_edge)
+		_add_edge(edge[0], edge[1], self.neighbors, self.reverse_neighbors, self.nodes, self.edges, l, self.interpretations_node, self.interpretations_edge, self.predicate_map_edge, self.num_ga, -1)
 
 	def add_node(self, node, labels):
 		# This function is useful for pyreason gym, called externally
@@ -648,11 +641,11 @@ class Interpretation:
 
 	def delete_edge(self, edge):
 		# This function is useful for pyreason gym, called externally
-		_delete_edge(edge, self.neighbors, self.reverse_neighbors, self.edges, self.interpretations_edge, self.predicate_map_edge)
+		_delete_edge(edge, self.neighbors, self.reverse_neighbors, self.edges, self.interpretations_edge, self.predicate_map_edge, self.num_ga)
 
 	def delete_node(self, node):
 		# This function is useful for pyreason gym, called externally
-		_delete_node(node, self.neighbors, self.reverse_neighbors, self.nodes, self.interpretations_node, self.predicate_map_node)
+		_delete_node(node, self.neighbors, self.reverse_neighbors, self.nodes, self.interpretations_node, self.predicate_map_node, self.num_ga)
 
 	def get_dict(self):
 		# This function can be called externally to retrieve a dict of the interpretation values
@@ -689,24 +682,30 @@ class Interpretation:
 
 		return interpretations
 
+	def get_final_num_ground_atoms(self):
+		"""
+		This function returns the number of ground atoms after the reasoning process, for the final timestep
+		:return: int: Number of ground atoms in the interpretation after reasoning
+		"""
+		ga_cnt = 0
+
+		for node in self.nodes:
+			for l in self.interpretations_node[node].world:
+				ga_cnt += 1
+		for edge in self.edges:
+			for l in self.interpretations_edge[edge].world:
+				ga_cnt += 1
+
+		return ga_cnt
+
 	def get_num_ground_atoms(self):
 		"""
 		This function returns the number of ground atoms after the reasoning process, for each timestep
-		:return: List: A list of number of ground atoms for each timestep
+		:return: list: Number of ground atoms in the interpretation after reasoning for each timestep
 		"""
-		num_ga = []
-
-		for t in range(self.time+1):
-			ga_cnt = 0
-			for node in self.nodes:
-				for l in self.interpretations_node[node].world:
-					ga_cnt += 1
-			for edge in self.edges:
-				for l in self.interpretations_edge[edge].world:
-					ga_cnt += 1
-			num_ga.append(ga_cnt)
-
-		return num_ga
+		if self.num_ga[-1] == 0:
+			self.num_ga.pop()
+		return self.num_ga
 
 	def query(self, query, return_bool=True) -> Union[bool, Tuple[float, float]]:
 		"""
@@ -751,7 +750,7 @@ class Interpretation:
 
 
 @numba.njit(cache=True)
-def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules):
+def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules, num_ga, t):
 	# Extract rule params
 	rule_type = rule.get_type()
 	head_variables = rule.get_head_variables()
@@ -1168,7 +1167,7 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 				if add_head_var_2_node_to_graph and head_var_2_grounding == head_var_2:
 					_add_node(head_var_2, neighbors, reverse_neighbors, nodes, interpretations_node)
 				if add_head_edge_to_graph and (head_var_1, head_var_2) == (head_var_1_grounding, head_var_2_grounding):
-					_add_edge(head_var_1, head_var_2, neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge)
+					_add_edge(head_var_1, head_var_2, neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
 
 				# For each grounding combination add a rule to be applied
 				# Only if all the clauses have valid groundings
@@ -2406,7 +2405,7 @@ def _satisfies_threshold(num_neigh, num_qualified_component, threshold):
 
 
 @numba.njit(cache=True)
-def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, mode, override=False):
+def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, num_ga, mode, override=False):
 	updated = False
 	# This is to prevent a key error in case the label is a specific label
 	try:
@@ -2417,6 +2416,7 @@ def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 		# Add label to world if it is not there
 		if l not in world.world:
 			world.world[l] = interval.closed(0, 1)
+			num_ga[t_cnt] += 1
 			if l in predicate_map:
 				predicate_map[l].append(comp)
 			else:
@@ -2511,7 +2511,7 @@ def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 
 
 @numba.njit(cache=True)
-def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, mode, override=False):
+def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, num_ga, mode, override=False):
 	updated = False
 	# This is to prevent a key error in case the label is a specific label
 	try:
@@ -2522,6 +2522,7 @@ def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 		# Add label to world if it is not there
 		if l not in world.world:
 			world.world[l] = interval.closed(0, 1)
+			num_ga[t_cnt] += 1
 			if l in predicate_map:
 				predicate_map[l].append(comp)
 			else:
@@ -2836,7 +2837,7 @@ def _add_node(node, neighbors, reverse_neighbors, nodes, interpretations_node):
 
 
 @numba.njit(cache=True)
-def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map):
+def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t):
 	# If not a node, add to list of nodes and initialize neighbors
 	if source not in nodes:
 		_add_node(source, neighbors, reverse_neighbors, nodes, interpretations_node)
@@ -2856,6 +2857,7 @@ def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, int
 		reverse_neighbors[target].append(source)
 		if l.value!='':
 			interpretations_edge[edge] = world.World(numba.typed.List([l]))
+			num_ga[t] += 1
 			if l in predicate_map:
 				predicate_map[l].append(edge)
 			else:
@@ -2866,26 +2868,28 @@ def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, int
 		if l not in interpretations_edge[edge].world and l.value!='':
 			new_edge = True
 			interpretations_edge[edge].world[l] = interval.closed(0, 1)
+			num_ga[t] += 1
 
 	return edge, new_edge
 
 
 @numba.njit(cache=True)
-def _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map):
+def _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t):
 	changes = 0
 	edges_added = numba.typed.List.empty_list(edge_type)
 	for source in sources:
 		for target in targets:
-			edge, new_edge = _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map)
+			edge, new_edge = _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t)
 			edges_added.append(edge)
 			changes = changes+1 if new_edge else changes
 	return edges_added, changes
 
 
 @numba.njit(cache=True)
-def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge, predicate_map):
+def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge, predicate_map, num_ga):
 	source, target = edge
 	edges.remove(edge)
+	num_ga[-1] -= len(interpretations_edge[edge].world)
 	del interpretations_edge[edge]
 	for l in predicate_map:
 		if edge in predicate_map[l]:
@@ -2895,8 +2899,9 @@ def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge
 
 
 @numba.njit(cache=True)
-def _delete_node(node, neighbors, reverse_neighbors, nodes, interpretations_node, predicate_map):
+def _delete_node(node, neighbors, reverse_neighbors, nodes, interpretations_node, predicate_map, num_ga):
 	nodes.remove(node)
+	num_ga[-1] -= len(interpretations_node[node].world)
 	del interpretations_node[node]
 	del neighbors[node]
 	del reverse_neighbors[node]
