+++ pyreason/scripts/interpretation/temp.py
@@ -1,4 +1,4 @@
-from typing import Union, Tuple
+from networkx.classes import edges
 
 import pyreason.scripts.numba_wrapper.numba_types.world_type as world
 import pyreason.scripts.numba_wrapper.numba_types.label_type as label
@@ -7,6 +7,7 @@ from pyreason.scripts.interpretation.interpretation_dict import InterpretationDi
 
 import numba
 from numba import objmode, prange
+import time
 
 
 # Types for the dictionaries
@@ -52,6 +53,8 @@ edges_to_be_added_type = numba.types.Tuple((numba.types.ListType(node_type), num
 
 
 class Interpretation:
+	available_labels_node = []
+	available_labels_edge = []
 	specific_node_labels = numba.typed.Dict.empty(key_type=label.label_type, value_type=numba.types.ListType(node_type))
 	specific_edge_labels = numba.typed.Dict.empty(key_type=label.label_type, value_type=numba.types.ListType(edge_type))
 
@@ -67,10 +70,6 @@ class Interpretation:
 		self.store_interpretation_changes = store_interpretation_changes
 		self.update_mode = update_mode
 		self.allow_ground_rules = allow_ground_rules
-
-		# Counter for number of ground atoms for each timestep, start with zero for the zeroth timestep
-		self.num_ga = numba.typed.List.empty_list(numba.types.int64)
-		self.num_ga.append(0)
 
 		# For reasoning and reasoning again (contains previous time and previous fp operation cnt)
 		self.time = 0
@@ -100,8 +99,18 @@ class Interpretation:
 		self.nodes.extend(numba.typed.List(self.graph.nodes()))
 		self.edges.extend(numba.typed.List(self.graph.edges()))
 
-		self.interpretations_node, self.predicate_map_node = self._init_interpretations_node(self.nodes, self.specific_node_labels, self.num_ga)
-		self.interpretations_edge, self.predicate_map_edge = self._init_interpretations_edge(self.edges, self.specific_edge_labels, self.num_ga)
+		# Make sure they are correct type
+		if len(self.available_labels_node)==0:
+			self.available_labels_node = numba.typed.List.empty_list(label.label_type)
+		else:
+			self.available_labels_node = numba.typed.List(self.available_labels_node)
+		if len(self.available_labels_edge)==0:
+			self.available_labels_edge = numba.typed.List.empty_list(label.label_type)
+		else:
+			self.available_labels_edge = numba.typed.List(self.available_labels_edge)
+
+		self.interpretations_node, self.predicate_map_node = self._init_interpretations_node(self.nodes, self.available_labels_node, self.specific_node_labels)
+		self.interpretations_edge, self.predicate_map_edge = self._init_interpretations_edge(self.edges, self.available_labels_edge, self.specific_edge_labels)
 
 		# Setup graph neighbors and reverse neighbors
 		self.neighbors = numba.typed.Dict.empty(key_type=node_type, value_type=numba.types.ListType(node_type))
@@ -130,19 +139,19 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True)
-	def _init_interpretations_node(nodes, specific_labels, num_ga):
+	def _init_interpretations_node(nodes, available_labels, specific_labels):
 		interpretations = numba.typed.Dict.empty(key_type=node_type, value_type=world.world_type)
 		predicate_map = numba.typed.Dict.empty(key_type=label.label_type, value_type=list_of_nodes)
-
-		# Initialize nodes
+		# General labels
 		for n in nodes:
-			interpretations[n] = world.World(numba.typed.List.empty_list(label.label_type))
-
+			interpretations[n] = world.World(available_labels)
 		# Specific labels
 		for l, ns in specific_labels.items():
 			for n in ns:
 				interpretations[n].world[l] = interval.closed(0.0, 1.0)
-				num_ga[0] += 1
+
+		for l in available_labels:
+			predicate_map[l] = numba.typed.List(nodes)
 
 		for l, ns in specific_labels.items():
 			predicate_map[l] = numba.typed.List(ns)
@@ -151,19 +160,19 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True)
-	def _init_interpretations_edge(edges, specific_labels, num_ga):
+	def _init_interpretations_edge(edges, available_labels, specific_labels):
 		interpretations = numba.typed.Dict.empty(key_type=edge_type, value_type=world.world_type)
 		predicate_map = numba.typed.Dict.empty(key_type=label.label_type, value_type=list_of_edges)
-
-		# Initialize edges
-		for n in edges:
-			interpretations[n] = world.World(numba.typed.List.empty_list(label.label_type))
-
+		# General labels
+		for e in edges:
+			interpretations[e] = world.World(available_labels)
 		# Specific labels
 		for l, es in specific_labels.items():
 			for e in es:
 				interpretations[e].world[l] = interval.closed(0.0, 1.0)
-				num_ga[0] += 1
+
+		for l in available_labels:
+			predicate_map[l] = numba.typed.List(edges)
 
 		for l, es in specific_labels.items():
 			predicate_map[l] = numba.typed.List(es)
@@ -213,7 +222,7 @@ class Interpretation:
 		return max_time
 
 	def _start_fp(self, rules, max_facts_time, verbose, again):
-		fp_cnt, t = self.reason(self.interpretations_node, self.interpretations_edge, self.predicate_map_node, self.predicate_map_edge, self.tmax, self.prev_reasoning_data, rules, self.nodes, self.edges, self.neighbors, self.reverse_neighbors, self.rules_to_be_applied_node, self.rules_to_be_applied_edge, self.edges_to_be_added_node_rule, self.edges_to_be_added_edge_rule, self.rules_to_be_applied_node_trace, self.rules_to_be_applied_edge_trace, self.facts_to_be_applied_node, self.facts_to_be_applied_edge, self.facts_to_be_applied_node_trace, self.facts_to_be_applied_edge_trace, self.ipl, self.rule_trace_node, self.rule_trace_edge, self.rule_trace_node_atoms, self.rule_trace_edge_atoms, self.reverse_graph, self.atom_trace, self.save_graph_attributes_to_rule_trace, self.canonical, self.inconsistency_check, self.store_interpretation_changes, self.update_mode, self.allow_ground_rules, max_facts_time, self.annotation_functions, self._convergence_mode, self._convergence_delta, self.num_ga, verbose, again)
+		fp_cnt, t = self.reason(self.interpretations_node, self.interpretations_edge, self.predicate_map_node, self.predicate_map_edge, self.tmax, self.prev_reasoning_data, rules, self.nodes, self.edges, self.neighbors, self.reverse_neighbors, self.rules_to_be_applied_node, self.rules_to_be_applied_edge, self.edges_to_be_added_node_rule, self.edges_to_be_added_edge_rule, self.rules_to_be_applied_node_trace, self.rules_to_be_applied_edge_trace, self.facts_to_be_applied_node, self.facts_to_be_applied_edge, self.facts_to_be_applied_node_trace, self.facts_to_be_applied_edge_trace, self.ipl, self.rule_trace_node, self.rule_trace_edge, self.rule_trace_node_atoms, self.rule_trace_edge_atoms, self.reverse_graph, self.atom_trace, self.save_graph_attributes_to_rule_trace, self.canonical, self.inconsistency_check, self.store_interpretation_changes, self.update_mode, self.allow_ground_rules, max_facts_time, self.annotation_functions, self._convergence_mode, self._convergence_delta, verbose, again)
 		self.time = t - 1
 		# If we need to reason again, store the next timestep to start from
 		self.prev_reasoning_data[0] = t
@@ -223,7 +232,7 @@ class Interpretation:
 
 	@staticmethod
 	@numba.njit(cache=True, parallel=False)
-	def reason(interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, tmax, prev_reasoning_data, rules, nodes, edges, neighbors, reverse_neighbors, rules_to_be_applied_node, rules_to_be_applied_edge, edges_to_be_added_node_rule, edges_to_be_added_edge_rule, rules_to_be_applied_node_trace, rules_to_be_applied_edge_trace, facts_to_be_applied_node, facts_to_be_applied_edge, facts_to_be_applied_node_trace, facts_to_be_applied_edge_trace, ipl, rule_trace_node, rule_trace_edge, rule_trace_node_atoms, rule_trace_edge_atoms, reverse_graph, atom_trace, save_graph_attributes_to_rule_trace, canonical, inconsistency_check, store_interpretation_changes, update_mode, allow_ground_rules, max_facts_time, annotation_functions, convergence_mode, convergence_delta, num_ga, verbose, again):
+	def reason(interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, tmax, prev_reasoning_data, rules, nodes, edges, neighbors, reverse_neighbors, rules_to_be_applied_node, rules_to_be_applied_edge, edges_to_be_added_node_rule, edges_to_be_added_edge_rule, rules_to_be_applied_node_trace, rules_to_be_applied_edge_trace, facts_to_be_applied_node, facts_to_be_applied_edge, facts_to_be_applied_node_trace, facts_to_be_applied_edge_trace, ipl, rule_trace_node, rule_trace_edge, rule_trace_node_atoms, rule_trace_edge_atoms, reverse_graph, atom_trace, save_graph_attributes_to_rule_trace, canonical, inconsistency_check, store_interpretation_changes, update_mode, allow_ground_rules, max_facts_time, annotation_functions, convergence_mode, convergence_delta, verbose, again):
 		t = prev_reasoning_data[0]
 		fp_cnt = prev_reasoning_data[1]
 		max_rules_time = 0
@@ -261,6 +270,8 @@ class Interpretation:
 
 			# Start by applying facts
 			# Nodes
+			with objmode(start='f8'):
+				start=time.time()
 			facts_to_be_applied_node_new.clear()
 			nodes_set = set(nodes)
 			for i in range(len(facts_to_be_applied_node)):
@@ -294,7 +305,7 @@ class Interpretation:
 						if check_consistent_node(interpretations_node, comp, (l, bnd)):
 							mode = 'graph-attribute-fact' if graph_attribute else 'fact'
 							override = True if update_mode == 'override' else False
-							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode=mode, override=override)
+							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode=mode, override=override)
 
 							update = u or update
 							# Update convergence params
@@ -308,7 +319,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_node(interpretations_node, comp, (l, bnd), ipl, t, fp_cnt, i, atom_trace, rule_trace_node, rule_trace_node_atoms, rules_to_be_applied_node_trace, facts_to_be_applied_node_trace, store_interpretation_changes, mode=mode)
 							else:
-								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode=mode, override=True)
+								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, i, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode=mode, override=True)
 
 								update = u or update
 								# Update convergence params
@@ -328,7 +339,13 @@ class Interpretation:
 			facts_to_be_applied_node[:] = facts_to_be_applied_node_new.copy()
 			facts_to_be_applied_node_new.clear()
 
+			with objmode():
+				print('Time taken for node facts:', time.time()-start, flush=True)
+
+
 			# Edges
+			with objmode(start='f8'):
+				start = time.time()
 			facts_to_be_applied_edge_new.clear()
 			edges_set = set(edges)
 			for i in range(len(facts_to_be_applied_edge)):
@@ -336,7 +353,7 @@ class Interpretation:
 					comp, l, bnd, static, graph_attribute = facts_to_be_applied_edge[i][1], facts_to_be_applied_edge[i][2], facts_to_be_applied_edge[i][3], facts_to_be_applied_edge[i][4], facts_to_be_applied_edge[i][5]
 					# If the component is not in the graph, add it
 					if comp not in edges_set:
-						_add_edge(comp[0], comp[1], neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
+						_add_edge(comp[0], comp[1], neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge)
 						edges_set.add(comp)
 
 					# Check if bnd is static. Then no need to update, just add to rule trace, check if graph attribute, and add ipl complement to rule trace as well
@@ -360,7 +377,7 @@ class Interpretation:
 						if check_consistent_edge(interpretations_edge, comp, (l, bnd)):
 							mode = 'graph-attribute-fact' if graph_attribute else 'fact'
 							override = True if update_mode == 'override' else False
-							u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode=mode, override=override)
+							u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode=mode, override=override)
 
 							update = u or update
 							# Update convergence params
@@ -374,7 +391,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_edge(interpretations_edge, comp, (l, bnd), ipl, t, fp_cnt, i, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode=mode)
 							else:
-								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode=mode, override=True)
+								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, i, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode=mode, override=True)
 
 								update = u or update
 								# Update convergence params
@@ -394,13 +411,20 @@ class Interpretation:
 			facts_to_be_applied_edge[:] = facts_to_be_applied_edge_new.copy()
 			facts_to_be_applied_edge_new.clear()
 
+			with objmode():
+				print('Time taken for edge facts:', time.time()-start, flush=True)
+
 			in_loop = True
 			while in_loop:
+				with objmode():
+					print('FP Iteration:', fp_cnt, flush=True)
 				# This will become true only if delta_t = 0 for some rule, otherwise we go to the next timestep
 				in_loop = False
 
 				# Apply the rules that need to be applied at this timestep
 				# Nodes
+				with objmode(start='f8'):
+					start = time.time()
 				rules_to_remove_idx.clear()
 				for idx, i in enumerate(rules_to_be_applied_node):
 					if i[0] == t:
@@ -408,7 +432,7 @@ class Interpretation:
 						# Check for inconsistencies
 						if check_consistent_node(interpretations_node, comp, (l, bnd)):
 							override = True if update_mode == 'override' else False
-							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
+							u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode='rule', override=override)
 
 							update = u or update
 							# Update convergence params
@@ -421,7 +445,7 @@ class Interpretation:
 							if inconsistency_check:
 								resolve_inconsistency_node(interpretations_node, comp, (l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_node, rule_trace_node_atoms, rules_to_be_applied_node_trace, facts_to_be_applied_node_trace, store_interpretation_changes, mode='rule')
 							else:
-								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
+								u, changes = _update_node(interpretations_node, predicate_map_node, comp, (l, bnd), ipl, rule_trace_node, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_node_trace, idx, facts_to_be_applied_node_trace, rule_trace_node_atoms, store_interpretation_changes, mode='rule', override=True)
 
 								update = u or update
 								# Update convergence params
@@ -439,13 +463,18 @@ class Interpretation:
 				if atom_trace:
 					rules_to_be_applied_node_trace[:] = numba.typed.List([rules_to_be_applied_node_trace[i] for i in range(len(rules_to_be_applied_node_trace)) if i not in rules_to_remove_idx])
 
+				with objmode():
+					print('Time taken for node rules:', time.time()-start, flush=True)
+
 				# Edges
+				with objmode(start='f8'):
+					start = time.time()
 				rules_to_remove_idx.clear()
 				for idx, i in enumerate(rules_to_be_applied_edge):
 					if i[0] == t:
 						comp, l, bnd, immediate, set_static = i[1], i[2], i[3], i[4], i[5]
 						sources, targets, edge_l = edges_to_be_added_edge_rule[idx]
-						edges_added, changes = _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, edge_l, interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
+						edges_added, changes = _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, edge_l, interpretations_node, interpretations_edge, predicate_map_edge)
 						changes_cnt += changes
 
 						# Update bound for newly added edges. Use bnd to update all edges if label is specified, else use bnd to update normally
@@ -455,7 +484,7 @@ class Interpretation:
 									continue
 								if check_consistent_edge(interpretations_edge, e, (edge_l, bnd)):
 									override = True if update_mode == 'override' else False
-									u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
+									u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=override)
 
 									update = u or update
 
@@ -469,7 +498,7 @@ class Interpretation:
 									if inconsistency_check:
 										resolve_inconsistency_edge(interpretations_edge, e, (edge_l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode='rule')
 									else:
-										u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
+										u, changes = _update_edge(interpretations_edge, predicate_map_edge, e, (edge_l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=True)
 
 										update = u or update
 
@@ -483,7 +512,7 @@ class Interpretation:
 							# Check for inconsistencies
 							if check_consistent_edge(interpretations_edge, comp, (l, bnd)):
 								override = True if update_mode == 'override' else False
-								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=override)
+								u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=override)
 
 								update = u or update
 								# Update convergence params
@@ -496,7 +525,7 @@ class Interpretation:
 								if inconsistency_check:
 									resolve_inconsistency_edge(interpretations_edge, comp, (l, bnd), ipl, t, fp_cnt, idx, atom_trace, rule_trace_edge, rule_trace_edge_atoms, rules_to_be_applied_edge_trace, facts_to_be_applied_edge_trace, store_interpretation_changes, mode='rule')
 								else:
-									u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, num_ga, mode='rule', override=True)
+									u, changes = _update_edge(interpretations_edge, predicate_map_edge, comp, (l, bnd), ipl, rule_trace_edge, fp_cnt, t, set_static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_edge_trace, idx, facts_to_be_applied_edge_trace, rule_trace_edge_atoms, store_interpretation_changes, mode='rule', override=True)
 
 									update = u or update
 									# Update convergence params
@@ -514,6 +543,9 @@ class Interpretation:
 				if atom_trace:
 					rules_to_be_applied_edge_trace[:] = numba.typed.List([rules_to_be_applied_edge_trace[i] for i in range(len(rules_to_be_applied_edge_trace)) if i not in rules_to_remove_idx])
 
+				with objmode():
+					print('Time taken for edge rules:', time.time()-start, flush=True)
+
 				# Fixed point
 				# if update or immediate_node_rule_fire or immediate_edge_rule_fire or immediate_rule_applied:
 				if update:
@@ -528,6 +560,8 @@ class Interpretation:
 						rules_to_be_applied_edge_trace_threadsafe = numba.typed.List([numba.typed.List.empty_list(rules_to_be_applied_trace_type) for _ in range(len(rules))])
 					edges_to_be_added_edge_rule_threadsafe = numba.typed.List([numba.typed.List.empty_list(edges_to_be_added_type) for _ in range(len(rules))])
 
+					with objmode(startrules='f8'):
+						startrules = time.time()
 					for i in prange(len(rules)):
 						rule = rules[i]
 						immediate_rule = rule.is_immediate_rule()
@@ -535,7 +569,11 @@ class Interpretation:
 						# Only go through if the rule can be applied within the given timesteps, or we're running until convergence
 						delta_t = rule.get_delta()
 						if t + delta_t <= tmax or tmax == -1 or again:
-							applicable_node_rules, applicable_edge_rules = _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules, num_ga, t)
+							with objmode(start='f8'):
+								start = time.time()
+							applicable_node_rules, applicable_edge_rules = _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules)
+							with objmode():
+								print('Time taken for grounding:', time.time()-start, flush=True)
 
 							# Loop through applicable rules and add them to the rules to be applied for later or next fp operation
 							for applicable_rule in applicable_node_rules:
@@ -580,6 +618,8 @@ class Interpretation:
 										in_loop = True
 										update = False
 
+					with objmode():
+						print('Time taken for ALL rule groundings:', time.time()-startrules, flush=True)
 					# Update lists after parallel run
 					for i in range(len(rules)):
 						if len(rules_to_be_applied_node_threadsafe[i]) > 0:
@@ -622,15 +662,14 @@ class Interpretation:
 					t += 1
 					break
 
-			# Increment t, update number of ground atoms
+			# Increment t
 			t += 1
-			num_ga.append(num_ga[-1])
 
 		return fp_cnt, t
 
 	def add_edge(self, edge, l):
 		# This function is useful for pyreason gym, called externally
-		_add_edge(edge[0], edge[1], self.neighbors, self.reverse_neighbors, self.nodes, self.edges, l, self.interpretations_node, self.interpretations_edge, self.predicate_map_edge, self.num_ga, -1)
+		_add_edge(edge[0], edge[1], self.neighbors, self.reverse_neighbors, self.nodes, self.edges, l, self.interpretations_node, self.interpretations_edge, self.predicate_map_edge)
 
 	def add_node(self, node, labels):
 		# This function is useful for pyreason gym, called externally
@@ -641,11 +680,11 @@ class Interpretation:
 
 	def delete_edge(self, edge):
 		# This function is useful for pyreason gym, called externally
-		_delete_edge(edge, self.neighbors, self.reverse_neighbors, self.edges, self.interpretations_edge, self.predicate_map_edge, self.num_ga)
+		_delete_edge(edge, self.neighbors, self.reverse_neighbors, self.edges, self.interpretations_edge, self.predicate_map_edge)
 
 	def delete_node(self, node):
 		# This function is useful for pyreason gym, called externally
-		_delete_node(node, self.neighbors, self.reverse_neighbors, self.nodes, self.interpretations_node, self.predicate_map_node, self.num_ga)
+		_delete_node(node, self.neighbors, self.reverse_neighbors, self.nodes, self.interpretations_node, self.predicate_map_node)
 
 	def get_dict(self):
 		# This function can be called externally to retrieve a dict of the interpretation values
@@ -682,43 +721,41 @@ class Interpretation:
 
 		return interpretations
 
-	def get_final_num_ground_atoms(self):
-		"""
-		This function returns the number of ground atoms after the reasoning process, for the final timestep
-		:return: int: Number of ground atoms in the interpretation after reasoning
-		"""
-		ga_cnt = 0
-
-		for node in self.nodes:
-			for l in self.interpretations_node[node].world:
-				ga_cnt += 1
-		for edge in self.edges:
-			for l in self.interpretations_edge[edge].world:
-				ga_cnt += 1
-
-		return ga_cnt
-
-	def get_num_ground_atoms(self):
-		"""
-		This function returns the number of ground atoms after the reasoning process, for each timestep
-		:return: list: Number of ground atoms in the interpretation after reasoning for each timestep
-		"""
-		if self.num_ga[-1] == 0:
-			self.num_ga.pop()
-		return self.num_ga
-
-	def query(self, query, return_bool=True) -> Union[bool, Tuple[float, float]]:
+	def query(self, query, return_bool=True):
 		"""
 		This function is used to query the graph after reasoning
-		:param query: A PyReason query object
+		:param query: The query string of for `pred(node)` or `pred(edge)` or `pred(node) : [l, u]`
 		:param return_bool: If True, returns boolean of query, else the bounds associated with it
 		:return: bool, or bounds
 		"""
+		# Parse the query
+		query = query.replace(' ', '')
 
-		comp_type = query.get_component_type()
-		component = query.get_component()
-		pred = query.get_predicate()
-		bnd = query.get_bounds()
+		if ':' in query:
+			pred_comp, bounds = query.split(':')
+			bounds = bounds.replace('[', '').replace(']', '')
+			l, u = bounds.split(',')
+			l, u = float(l), float(u)
+		else:
+			if query[0] == '~':
+				pred_comp = query[1:]
+				l, u = 0, 0
+			else:
+				pred_comp = query
+				l, u = 1, 1
+
+		bnd = interval.closed(l, u)
+
+		# Split predicate and component
+		idx = pred_comp.find('(')
+		pred = label.Label(pred_comp[:idx])
+		component = pred_comp[idx + 1:-1]
+
+		if ',' in component:
+			component = tuple(component.split(','))
+			comp_type = 'edge'
+		else:
+			comp_type = 'node'
 
 		# Check if the component exists
 		if comp_type == 'node':
@@ -750,7 +787,7 @@ class Interpretation:
 
 
 @numba.njit(cache=True)
-def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules, num_ga, t):
+def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map_node, predicate_map_edge, nodes, edges, neighbors, reverse_neighbors, atom_trace, allow_ground_rules):
 	# Extract rule params
 	rule_type = rule.get_type()
 	head_variables = rule.get_head_variables()
@@ -799,12 +836,18 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 
 			# Get subset of nodes that can be used to ground the variable
 			# If we allow ground atoms, we can use the nodes directly
+			with objmode(start='f8'):
+				start = time.time()
 			if allow_ground_rules and clause_var_1 in nodes_set:
 				grounding = numba.typed.List([clause_var_1])
 			else:
-				grounding = get_rule_node_clause_grounding(clause_var_1, groundings, predicate_map_node, clause_label, nodes)
+				grounding = get_rule_node_clause_grounding(clause_var_1, groundings, predicate_map_node, clause_label)
+			with objmode():
+				print('__ get node groundings:', time.time()-start, flush=True)
 
 			# Narrow subset based on predicate
+			with objmode(start='f8'):
+				start = time.time()
 			qualified_groundings = get_qualified_node_groundings(interpretations_node, grounding, clause_label, clause_bnd)
 			groundings[clause_var_1] = qualified_groundings
 			qualified_groundings_set = set(qualified_groundings)
@@ -813,6 +856,9 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 					groundings_edges[(c1, c2)] = numba.typed.List([e for e in groundings_edges[(c1, c2)] if e[0] in qualified_groundings_set])
 				if c2 == clause_var_1:
 					groundings_edges[(c1, c2)] = numba.typed.List([e for e in groundings_edges[(c1, c2)] if e[1] in qualified_groundings_set])
+
+			with objmode():
+				print('__ get node qualified groundings:', time.time()-start, flush=True)
 
 			# Check satisfaction of those nodes wrt the threshold
 			# Only check satisfaction if the default threshold is used. This saves us from grounding the rest of the rule
@@ -826,13 +872,23 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 
 			# Get subset of edges that can be used to ground the variables
 			# If we allow ground atoms, we can use the nodes directly
+			with objmode(start='f8'):
+				start = time.time()
 			if allow_ground_rules and (clause_var_1, clause_var_2) in edges_set:
 				grounding = numba.typed.List([(clause_var_1, clause_var_2)])
 			else:
-				grounding = get_rule_edge_clause_grounding(clause_var_1, clause_var_2, groundings, groundings_edges, neighbors, reverse_neighbors, predicate_map_edge, clause_label, edges)
+				grounding = get_rule_edge_clause_grounding(clause_var_1, clause_var_2, groundings, groundings_edges, neighbors, reverse_neighbors, predicate_map_edge, clause_label)
 
+			with objmode():
+				print('__ get edge groundings:', time.time()-start, flush=True)
+
+			with objmode(start='f8'):
+				start = time.time()
 			# Narrow subset based on predicate (save the edges that are qualified to use for finding future groundings faster)
 			qualified_groundings = get_qualified_edge_groundings(interpretations_edge, grounding, clause_label, clause_bnd)
+
+			with objmode():
+				print('__ get edge qualified groundings:', time.time()-start, flush=True)
 
 			# Check satisfaction of those edges wrt the threshold
 			# Only check satisfaction if the default threshold is used. This saves us from grounding the rest of the rule
@@ -841,6 +897,8 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 			satisfaction = check_edge_grounding_threshold_satisfaction(interpretations_edge, grounding, qualified_groundings, clause_label, thresholds[i]) and satisfaction
 
 			# Update the groundings
+			with objmode(start='f8'):
+				start = time.time()
 			groundings[clause_var_1] = numba.typed.List.empty_list(node_type)
 			groundings[clause_var_2] = numba.typed.List.empty_list(node_type)
 			groundings_clause_1_set = set(groundings[clause_var_1])
@@ -867,13 +925,20 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 			elif clause_var_1 not in dependency_graph_reverse_neighbors[clause_var_2]:
 				dependency_graph_reverse_neighbors[clause_var_2].append(clause_var_1)
 
+			with objmode():
+				print('__ updated dependency graph and groundings:', time.time()-start, flush=True)
+
 		# This is a comparison clause
 		else:
 			pass
 
 		# Refine the subsets based on any updates
 		if satisfaction:
+			with objmode(start='f8'):
+				start = time.time()
 			refine_groundings(clause_variables, groundings, groundings_edges, dependency_graph_neighbors, dependency_graph_reverse_neighbors)
+			with objmode():
+				print('__ refined groundings:', time.time()-start, flush=True)
 
 		# If satisfaction is false, break
 		if not satisfaction:
@@ -891,6 +956,8 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 		if rule_type == 'node':
 			# Loop through all the head variable groundings and add it to the rules to be applied
 			# Loop through the clauses and add appropriate trace data and annotations
+			with objmode(start='f8'):
+				start = time.time()
 
 			# If there is no grounding for head_var_1, we treat it as a ground atom and add it to the graph
 			head_var_1_in_nodes = head_var_1 in nodes
@@ -978,9 +1045,14 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 				# For each grounding add a rule to be applied
 				applicable_rules_node.append((head_grounding, annotations, qualified_nodes, qualified_edges, edges_to_be_added))
 
+			with objmode():
+				print('__ rule fire node head grounding:', time.time()-start, flush=True)
+
 		elif rule_type == 'edge':
 			head_var_1 = head_variables[0]
 			head_var_2 = head_variables[1]
+			with objmode(start='f8'):
+				start = time.time()
 
 			# If there is no grounding for head_var_1 or head_var_2, we treat it as a ground atom and add it to the graph
 			head_var_1_in_nodes = head_var_1 in nodes
@@ -1167,13 +1239,16 @@ def _ground_rule(rule, interpretations_node, interpretations_edge, predicate_map
 				if add_head_var_2_node_to_graph and head_var_2_grounding == head_var_2:
 					_add_node(head_var_2, neighbors, reverse_neighbors, nodes, interpretations_node)
 				if add_head_edge_to_graph and (head_var_1, head_var_2) == (head_var_1_grounding, head_var_2_grounding):
-					_add_edge(head_var_1, head_var_2, neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge, num_ga, t)
+					_add_edge(head_var_1, head_var_2, neighbors, reverse_neighbors, nodes, edges, label.Label(''), interpretations_node, interpretations_edge, predicate_map_edge)
 
 				# For each grounding combination add a rule to be applied
 				# Only if all the clauses have valid groundings
 				# if satisfaction:
 				e = (head_var_1_grounding, head_var_2_grounding)
 				applicable_rules_edge.append((e, annotations, qualified_nodes, qualified_edges, edges_to_be_added))
+
+			with objmode():
+				print('__ rule fire edge head grounding:', time.time()-start, flush=True)
 
 	# Return the applicable rules
 	return applicable_rules_node, applicable_rules_edge
@@ -1977,17 +2052,14 @@ def check_edge_clause_satisfaction(interpretations_edge, subsets, subset_source,
 
 
 @numba.njit(cache=True)
-def get_rule_node_clause_grounding(clause_var_1, groundings, predicate_map, l, nodes):
+def get_rule_node_clause_grounding(clause_var_1, groundings, predicate_map, l):
 	# The groundings for a node clause can be either a previous grounding or all possible nodes
-	if l in predicate_map:
-		grounding = predicate_map[l] if clause_var_1 not in groundings else groundings[clause_var_1]
-	else:
-		grounding = nodes if clause_var_1 not in groundings else groundings[clause_var_1]
+	grounding = predicate_map[l] if clause_var_1 not in groundings else groundings[clause_var_1]
 	return grounding
 
 
 @numba.njit(cache=True)
-def get_rule_edge_clause_grounding(clause_var_1, clause_var_2, groundings, groundings_edges, neighbors, reverse_neighbors, predicate_map, l, edges):
+def get_rule_edge_clause_grounding(clause_var_1, clause_var_2, groundings, groundings_edges, neighbors, reverse_neighbors, predicate_map, l):
 	# There are 4 cases for predicate(Y,Z):
 	# 1. Both predicate variables Y and Z have not been encountered before
 	# 2. The source variable Y has not been encountered before but the target variable Z has
@@ -1998,10 +2070,7 @@ def get_rule_edge_clause_grounding(clause_var_1, clause_var_2, groundings, groun
 	# Case 1:
 	# We replace Y by all nodes and Z by the neighbors of each of these nodes
 	if clause_var_1 not in groundings and clause_var_2 not in groundings:
-		if l in predicate_map:
-			edge_groundings = predicate_map[l]
-		else:
-			edge_groundings = edges
+		edge_groundings = predicate_map[l]
 
 	# Case 2:
 	# We replace Y by the sources of Z
@@ -2405,7 +2474,7 @@ def _satisfies_threshold(num_neigh, num_qualified_component, threshold):
 
 
 @numba.njit(cache=True)
-def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, num_ga, mode, override=False):
+def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, mode, override=False):
 	updated = False
 	# This is to prevent a key error in case the label is a specific label
 	try:
@@ -2416,7 +2485,6 @@ def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 		# Add label to world if it is not there
 		if l not in world.world:
 			world.world[l] = interval.closed(0, 1)
-			num_ga[t_cnt] += 1
 			if l in predicate_map:
 				predicate_map[l].append(comp)
 			else:
@@ -2511,7 +2579,7 @@ def _update_node(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 
 
 @numba.njit(cache=True)
-def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, num_ga, mode, override=False):
+def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_cnt, t_cnt, static, convergence_mode, atom_trace, save_graph_attributes_to_rule_trace, rules_to_be_applied_trace, idx, facts_to_be_applied_trace, rule_trace_atoms, store_interpretation_changes, mode, override=False):
 	updated = False
 	# This is to prevent a key error in case the label is a specific label
 	try:
@@ -2522,7 +2590,6 @@ def _update_edge(interpretations, predicate_map, comp, na, ipl, rule_trace, fp_c
 		# Add label to world if it is not there
 		if l not in world.world:
 			world.world[l] = interval.closed(0, 1)
-			num_ga[t_cnt] += 1
 			if l in predicate_map:
 				predicate_map[l].append(comp)
 			else:
@@ -2764,12 +2831,10 @@ def resolve_inconsistency_node(interpretations, comp, na, ipl, t_cnt, fp_cnt, id
 	world = interpretations[comp]
 	if store_interpretation_changes:
 		rule_trace.append((numba.types.uint16(t_cnt), numba.types.uint16(fp_cnt), comp, na[0], interval.closed(0,1)))
-		if mode == 'fact' or mode == 'graph-attribute-fact' and atom_trace:
+		if mode == 'fact' or mode == 'graph-attribute-fact':
 			name = facts_to_be_applied_trace[idx]
-		elif mode == 'rule' and atom_trace:
+		elif mode == 'rule':
 			name = rules_to_be_applied_trace[idx][2]
-		else:
-			name = '-'
 		if atom_trace:
 			_update_rule_trace(rule_trace_atoms, numba.typed.List.empty_list(numba.typed.List.empty_list(node_type)), numba.typed.List.empty_list(numba.typed.List.empty_list(edge_type)), world.world[na[0]], f'Inconsistency due to {name}')
 	# Resolve inconsistency and set static
@@ -2799,12 +2864,10 @@ def resolve_inconsistency_edge(interpretations, comp, na, ipl, t_cnt, fp_cnt, id
 	w = interpretations[comp]
 	if store_interpretation_changes:
 		rule_trace.append((numba.types.uint16(t_cnt), numba.types.uint16(fp_cnt), comp, na[0], interval.closed(0,1)))
-		if mode == 'fact' or mode == 'graph-attribute-fact' and atom_trace:
+		if mode == 'fact' or mode == 'graph-attribute-fact':
 			name = facts_to_be_applied_trace[idx]
-		elif mode == 'rule' and atom_trace:
+		elif mode == 'rule':
 			name = rules_to_be_applied_trace[idx][2]
-		else:
-			name = '-'
 		if atom_trace:
 			_update_rule_trace(rule_trace_atoms, numba.typed.List.empty_list(numba.typed.List.empty_list(node_type)), numba.typed.List.empty_list(numba.typed.List.empty_list(edge_type)), w.world[na[0]], f'Inconsistency due to {name}')
 	# Resolve inconsistency and set static
@@ -2837,7 +2900,7 @@ def _add_node(node, neighbors, reverse_neighbors, nodes, interpretations_node):
 
 
 @numba.njit(cache=True)
-def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t):
+def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map):
 	# If not a node, add to list of nodes and initialize neighbors
 	if source not in nodes:
 		_add_node(source, neighbors, reverse_neighbors, nodes, interpretations_node)
@@ -2857,7 +2920,6 @@ def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, int
 		reverse_neighbors[target].append(source)
 		if l.value!='':
 			interpretations_edge[edge] = world.World(numba.typed.List([l]))
-			num_ga[t] += 1
 			if l in predicate_map:
 				predicate_map[l].append(edge)
 			else:
@@ -2868,28 +2930,26 @@ def _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, int
 		if l not in interpretations_edge[edge].world and l.value!='':
 			new_edge = True
 			interpretations_edge[edge].world[l] = interval.closed(0, 1)
-			num_ga[t] += 1
 
 	return edge, new_edge
 
 
 @numba.njit(cache=True)
-def _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t):
+def _add_edges(sources, targets, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map):
 	changes = 0
 	edges_added = numba.typed.List.empty_list(edge_type)
 	for source in sources:
 		for target in targets:
-			edge, new_edge = _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map, num_ga, t)
+			edge, new_edge = _add_edge(source, target, neighbors, reverse_neighbors, nodes, edges, l, interpretations_node, interpretations_edge, predicate_map)
 			edges_added.append(edge)
 			changes = changes+1 if new_edge else changes
 	return edges_added, changes
 
 
 @numba.njit(cache=True)
-def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge, predicate_map, num_ga):
+def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge, predicate_map):
 	source, target = edge
 	edges.remove(edge)
-	num_ga[-1] -= len(interpretations_edge[edge].world)
 	del interpretations_edge[edge]
 	for l in predicate_map:
 		if edge in predicate_map[l]:
@@ -2899,9 +2959,8 @@ def _delete_edge(edge, neighbors, reverse_neighbors, edges, interpretations_edge
 
 
 @numba.njit(cache=True)
-def _delete_node(node, neighbors, reverse_neighbors, nodes, interpretations_node, predicate_map, num_ga):
+def _delete_node(node, neighbors, reverse_neighbors, nodes, interpretations_node, predicate_map):
 	nodes.remove(node)
-	num_ga[-1] -= len(interpretations_node[node].world)
 	del interpretations_node[node]
 	del neighbors[node]
 	del reverse_neighbors[node]
